Create webcrawler
This is a simple web crawler that crawls through all the pages up to the depth of 3. It performs focused and focused crawling for the keyword being passed as arguments. The initial seed page is also passed as an argument to the file.

I have used BFS algorithm since it is the most efficient for this implementation of web crawler.